Move mixins to tasks app
- Move StatusTrackingMixin and TimestampMixin to apps/tasks/mixins.py
- Update apps/tasks/models.py to import from local mixins
- Update test imports
- Delete apps/core/mixins.py
